+++ cpp09/ex01/RPN.hpp
@@ -8,7 +8,7 @@
 
 class RPN {
 	private:
-		std::stack<int> _stack;
+		std::stack<float> _stack;
 		std::string _str;
 		RPN();
